+++ src/models/Settings.php
@@ -11,14 +11,11 @@ use craft\base\Model;
 class Settings extends Model {
     public bool $toolbarEnabled = true;
 
+    public bool $startMenuEnabled = true;
     public bool $globalSearchEnabled = true;
     public bool $actionButtonsEnabled = true;
     public bool $seoMenuEnabled = true;
-    public bool $seoChecklistEnabled = true;
-    public bool $seoSharePreviewsEnabled = true;
-    public bool $seoImagesEnabled = true;
-    public bool $seoHeadingsEnabled = true;
-    public bool $seoStructuredDataEnabled = true;
+    public bool $helpMenuEnabled = true;
     public array $customWidgets = [];
 
     public string $company = '';